Events and more logic

+++ src/main/kotlin/com/martmists/hackgame/server/ServerCommands.kt
@@ -9,23 +9,24 @@ import com.martmists.hackgame.server.entities.ServerCommandSource
 import com.martmists.hackgame.common.entities.TextColor
 import com.mojang.brigadier.CommandDispatcher
 import com.mojang.brigadier.arguments.StringArgumentType
+import kotlinx.serialization.ExperimentalSerializationApi
 
 object ServerCommands {
     fun initialize(dispatcher: CommandDispatcher<ServerCommandSource>) {
         CommandBuilder.builder(dispatcher) {
             command("money") {
                 executes {
-                    BuiltinPackets.FEEDBACK_S2C.send(FeedbackPacket("${TextColor.ANSI.YELLOW}${it.source.currentHost.money}$${TextColor.ANSI.WHITE}"), it.source.connection)
+                    BuiltinPackets.FEEDBACK_S2C.send(FeedbackPacket("${TextColor.ANSI.YELLOW}${it.source.currentHost.money}$"), it.source.connection)
                 }
             }
 
-            command("ls", "dir") {
+            command("ls") {
                 executes {
                     val tree = it.source.currentHost.filesystem.generateView()
                     BuiltinPackets.FEEDBACK_S2C.send(FeedbackPacket(tree), it.source.connection)
                 }
 
-                argument("path", StringArgumentType.string()) {
+                argument("path", StringArgumentType.greedyString()) {
                     executes {
                         val root = it.source.currentHost.filesystem.getDirByPath(StringArgumentType.getString(it, "path"))
                         val tree = root.generateView()
@@ -35,7 +36,7 @@ object ServerCommands {
             }
 
             command("cat") {
-                argument("path", StringArgumentType.string()) {
+                argument("path", StringArgumentType.greedyString()) {
                     executes {
                         val file = it.source.currentHost.filesystem.getFileByPath(StringArgumentType.getString(it, "path"))
                         BuiltinPackets.FEEDBACK_S2C.send(FeedbackPacket(file.contents), it.source.connection)
@@ -44,9 +45,13 @@ object ServerCommands {
             }
 
             command("rm") {
-                argument("path", StringArgumentType.string()) {
+                argument("path", StringArgumentType.greedyString()) {
                     executes {
                         val parts = StringArgumentType.getString(it, "path").split("/").toMutableList()
+                        if (parts.first() == "bin") {
+                            BuiltinPackets.FEEDBACK_S2C.send(FeedbackPacket("${TextColor.ANSI.RED}Cannot delete files in bin"), it.source.connection)
+                            return@executes
+                        }
                         val file = parts.removeLast()
                         val dir = it.source.currentHost.filesystem.getDirByPath(parts.joinToString("/"))
                         dir.removeFile(file)
@@ -55,9 +60,13 @@ object ServerCommands {
             }
 
             command("rmdir") {
-                argument("path", StringArgumentType.string()) {
+                argument("path", StringArgumentType.greedyString()) {
                     executes {
                         val parts = StringArgumentType.getString(it, "path").split("/").toMutableList()
+                        if (parts.first() == "bin") {
+                            BuiltinPackets.FEEDBACK_S2C.send(FeedbackPacket("${TextColor.ANSI.RED}Cannot delete folders in bin"), it.source.connection)
+                            return@executes
+                        }
                         val dirname = parts.removeLast()
                         val dir = it.source.currentHost.filesystem.getDirByPath(parts.joinToString("/"))
                         dir.removeDir(dirname)
@@ -76,7 +85,7 @@ object ServerCommands {
             command("disconnect") {
                 executes {
                     if (it.source.currentHost == it.source.ownHost) {
-                        BuiltinPackets.FEEDBACK_S2C.send(FeedbackPacket("${TextColor.ANSI.RED}ERROR: Cannot disconnect from root${TextColor.ANSI.WHITE}"), it.source.connection)
+                        BuiltinPackets.FEEDBACK_S2C.send(FeedbackPacket("${TextColor.ANSI.RED}ERROR: Cannot disconnect from root"), it.source.connection)
                     } else {
                         val last = it.source.session.connectChain.pop()
                         it.source.session.currentIP = it.source.currentHost.ip
@@ -101,7 +110,7 @@ object ServerCommands {
 
             command("commands") {
                 executes {
-                    val commands = dispatcher.root.children.joinToString("\n") { c -> "- ${c.name}" }
+                    val commands = dispatcher.root.children.joinToString("\n") { c -> "- ${TextColor.ANSI.GREEN}${c.name}${TextColor.ANSI.WHITE}" }
                     BuiltinPackets.FEEDBACK_S2C.send(FeedbackPacket("Commands:\n$commands"), it.source.connection)
                 }
 
@@ -112,6 +121,30 @@ object ServerCommands {
                         // Error message if command is null
                         // Command usage if command is not null
                     }
+                }
+            }
+
+            command("software") {
+                executes {
+                    BuiltinPackets.FEEDBACK_S2C.send(FeedbackPacket("${TextColor.ANSI.WHITE}Usage: software [install|uninstall|list]"), it.source.connection)
+                }
+
+                literal("install") {
+                    // Move to bin folder
+                    argument("software", StringArgumentType.string()) {
+                        executes {
+                            // If file exists, install, add to software and move
+                            // If file not exists, error
+                        }
+                    }
+                }
+
+                literal("uninstall") {
+                    // move to software folder
+                }
+
+                literal("list") {
+                    // list installed software
                 }
             }
         }
